Creation of the path get car by id

+++ src/main/java/br/com/luisergio/cardealership/business/CarBusiness.java
@@ -8,5 +8,5 @@ public interface CarBusiness {
 
     List<CarDto> getAll();
 
-    CarDto getById(Integer id);
+    CarDto getById(Long id);
 }
